fix registration and account recovery
On Vue 3, ref names and v-model names are both assigned to the scope of the component,
unlike in Vue 2, where refs were only registered on vm.$refs. This means if you
have a <input v-model="blah" ref="blah"> the ref will override the model and thus
the model will not work. This is, of course, undocumented in the migration guide 🙄