Added Plausible analytics
This re-enables a simple performance metric without compromising the
privacy of our visitors. Plausible works without cookies and does not
perform user tracking, so it does not need a cookie banner.

I can cover the service cost for now since I'm paying anyway for the
mapeditor.org statistics. If this becomes a problem in the future we can
also choose to self-host it.